Added Edit Book Feature

+++ server/deleteUser.php
@@ -1,4 +1,9 @@
 <?php
+error_reporting(-1); // reports all errors
+ini_set("display_errors", "1"); // shows all errors
+ini_set("log_errors", 1);
+ini_set("error_log", "/tmp/php-error.log");
+
 header("Access-Control-Allow-Origin: *");
 header("Access-Control-Allow-Headers: access");
 header("Access-Control-Allow-Methods: POST");
@@ -13,15 +18,33 @@ if ($_SERVER['REQUEST_METHOD'] === 'OPTIONS') {
 $data = json_decode(file_get_contents("php://input"));
 if (isset($data->id) && is_numeric($data->id)) {
     $delID = $data->id;
-    $deleteUser = mysqli_query($conn, "DELETE FROM `users` WHERE `id`='$delID'");
-    if ($deleteUser) {
-        header("HTTP/1.1 200 OK");
-        echo json_encode(["msg"=>"User Deleted"]);
-    } else {
-        header("HTTP/1.1 500 NOT DELETED");
-        echo json_encode(["msg"=>"User Not Deleted!"]);
+    $getUserCurrentBookId = mysqli_query($conn, "SELECT `currently_issued_bookid` FROM `users` WHERE `id`='$delID'");
+    $book_arr = mysqli_fetch_all($getUserCurrentBookId, MYSQLI_ASSOC);
+    print_r($book_arr) ;
+    print_r($getUserCurrentBookId) ;
+    $temp = $book_arr[0]['currently_issued_bookid'];
+    if($temp != NULL){
+        
+        $updateBookCount = mysqli_query($conn,"UPDATE books SET `available_count`=`available_count`+1 WHERE `id`='$temp'");
+        if($updateBookCount) {
+            $deleteUser = mysqli_query($conn, "DELETE FROM `users` WHERE `id`='$delID'");
+    
+            if ($deleteUser) {
+                header("HTTP/1.1 200 OK");
+                echo json_encode(["msg"=>"User Deleted"]);
+            } else {
+                header("HTTP/1.1 500 NOT DELETED");
+                echo json_encode(["msg"=>"User Not Deleted!"]);
+            }
+            
+        }
+        else {
+            header("HTTP/1.1 401 RESOURCE NOT FOUND");
+            echo json_encode(["msg"=>"Update Failed!"]);
+        }
     }
+    
 } else {
-    header("HTTP/1.1 403 RESOURCE NOT FOUND");
+    header("HTTP/1.1 401 RESOURCE NOT FOUND");
     echo json_encode(["msg"=>"User Not Found!"]);
 }
